CDDL works fully for JSON and CBOR; remove UCCS reference
An unfortunately large commit the main purpose of which truly makes the CDDL work for both JSON and CBOR encoding. With this there are both CBOR and JSON examples that validate against the CDDL in the document. This also boosts confidence the CDDL a lot.

This PR also removes any reference to UCCS and discussion of UJCS. In their stead there is a socket into which new EAT formats like UCCS and UJCS should plug in to.

This also sorts out all the external CDDL. Some is published in the document marked as a duplicate of some external CDDL or to-be published CDDL (Claims-Set in particular).

The document build system fetches some external CDDL (SUIT and CoSWID) and has stubs for other external CDDL (COSE) to enable the full validation of the examples.

Note that not every shred of CDDL has an example. Some of it is not cross-checked by validation against some JSON or CBOR.

The detailed list of commits:

* first go at using JC<> template from uccs

* CBOR submods are validating; add some missing files to repo

* Update external cddl

* improve JC<> usage for debug status and oemid

* Added Claims-Set CDDL appendix, doc compiles...

* Clean up download of external cddl

* fix capitalization of claims-set-claim

* lots of work on enumerated types

* Check point work on CDDL for JSON and CBOR

* A few file renames

* more file renames

* split examples in to payload and token

* fix file name

* fix another file name

* new better handling of unassigned CBOR labels and tag numbers

* Fully remove all UCCS, a few other fix ups

* Remove UCCS, re order sections, ...

- UCCS is removed
- the JC<> generic is used where CDDL varies between JSON and CBOR
- New section on top-level of token
- Put claims into four separate sub sections
- Lots of reorganizing of the CDDL


* first go at using JC<> template from uccs

* CBOR submods are validating; add some missing files to repo

* Update external cddl

* improve JC<> usage for debug status and oemid

* Added Claims-Set CDDL appendix, doc compiles...

* Clean up download of external cddl

* fix capitalization of claims-set-claim

* lots of work on enumerated types

* Check point work on CDDL for JSON and CBOR

* A few file renames

* more file renames

* split examples in to payload and token

* fix file name

* fix another file name

* new better handling of unassigned CBOR labels and tag numbers

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>

* more removal of UCCS

* remove UJCS

* Clean up odds and ends

* add valid CWT token example now that UCCS examples can't be used

* Put a valid CWT in the DEB now that UCCS can't be used

* add a JWT submod example

* CDDL bug fixes; CBOR and JSON examples validating

* Forgot a file

* All examples are validating

* Fix OIDs for JSON

* Fix use of uri common data type

* Fix time and URI types

* Final clean up of time types

* fill out more of the changed-since

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>