Clarify the security level claimed  (#162)
The text that was there prior to this change was not good the useful intent of this claim. It was pulling people to claim the highest level they achieved which is not useful to the RP.

(We talked about not merging this, but on second thought, I think the previous text was too misleading. I hope the new text that clearly says this is only a rough characterization of security helps side step the issue of having to provide or perform detailed analysis of the entity).

* clarify min/max security level

* wording improvement

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>