Change how detached digests are identified in JSON tokens
This is a protocol change to the way detached digests are identified submodules in JSON tokens. Previously there was some overloading of the array in a JSON-encoded submodule where it could be either a selector or a detached digest. The detached digest is now pushed one layer down and the overloading is removed.

This also has some renaming of the CDDL elements that should be more clear and some combining of CDDL files that also should help readability.


* initial changes towards modified submodule definition

* - remove reference to deleted submods.cddl file and delete the file
- change group sockets to type sockets in eat-json.cddl and uccs.cddl
- bring Nested-Token type back (since it is used in bundles)
- text changes mostly related to clean up references to abstract nested tokens vs references to Nested-Token

* add socket group to terminology section

* change JSON-Nested-Token-Or-Detached-Digest to JSON-Selector

* move contents of nested-token-xxx files into corresponding submods-xxx file. refactor prose and definitions such that Nested-Token is only used in Detached-EAT-Bundle. Use JSON-Selector for JSON embedded in CBOR, with caveat that DIGEST option must not be used.