require presence of hwmodel, swname claims (#376)
This change requires the presence of the "oemid" claim if the "hwmodel" is present and the present of the "hwmodel" claim is the "hwversion" claim is present. Same for "swversion" and "swname".

At first I was thinking to be flexible and generous to allow variability here, but now I'm thinking it is better to be more specific and strict. It will make the token ecosystem more useful to the consumer.


* require presence of hwmodel, swname claims

* wording

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>