Fix/update examples and CDDL validation (#450)
This updates the examples with the IANA-assigned labels and tags.

This fixes some errors in the examples, mostly missing commas.

CDDL validation is passing again. This validates consistency between the normative CDDL and the examples.

There is a change to the normative CDDL to compensate for a bug in the cddl tool. RFC 8610 says that there is a problem with generics nesting in the cddl tool. This is presumed to be the cause of the problem. The solution is to provide a non-nested version of the JC<> generic for use in contexts where the cddl validation gets confused. (It didn't produce an error, it just didn't validate correct input; the solution was found by extensive trial and error).


* Fix/update examples and CDDL validation

* Better full CWT example; formatting

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>